+++ Rebus.AzureStorage.Tests/Transport/AzureStorageQueuesTransportFactory.cs
@@ -1,8 +1,5 @@
 ﻿using System;
 using System.Collections.Concurrent;
-using System.Configuration;
-using System.IO;
-using Microsoft.WindowsAzure.Storage;
 using Rebus.AzureStorage.Transport;
 using Rebus.Logging;
 using Rebus.Tests.Contracts.Transports;
@@ -23,8 +20,7 @@ namespace Rebus.AzureStorage.Tests.Transport
         {
             if (inputQueueAddress == null)
             {
-                var storageAccount = CloudStorageAccount.Parse(ConnectionString);
-                var transport = new AzureStorageQueuesTransport(storageAccount, null, new ConsoleLoggerFactory(false));
+                var transport = new AzureStorageQueuesTransport(AzureConfig.StorageAccount, null, new ConsoleLoggerFactory(false));
 
                 transport.Initialize();
 
@@ -33,8 +29,7 @@ namespace Rebus.AzureStorage.Tests.Transport
 
             return _transports.GetOrAdd(inputQueueAddress, address =>
             {
-                var storageAccount = CloudStorageAccount.Parse(ConnectionString);
-                var transport = new AzureStorageQueuesTransport(storageAccount, inputQueueAddress, new ConsoleLoggerFactory(false));
+                var transport = new AzureStorageQueuesTransport(AzureConfig.StorageAccount, inputQueueAddress, new ConsoleLoggerFactory(false));
 
                 transport.PurgeInputQueue();
 
@@ -42,42 +37,6 @@ namespace Rebus.AzureStorage.Tests.Transport
 
                 return transport;
             });
-        }
-
-        public static string ConnectionString => ConnectionStringFromFileOrNull(Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "..", "..", "azure_storage_connection_string.txt"))
-                                                 ?? ConnectionStringFromEnvironmentVariable("rebus2_storage_connection_string")
-                                                 ?? Throw("Could not find Azure Storage connection string!");
-
-        static string ConnectionStringFromFileOrNull(string filePath)
-        {
-            if (!File.Exists(filePath))
-            {
-                Console.WriteLine("Could not find file {0}", filePath);
-                return null;
-            }
-
-            Console.WriteLine("Using Azure Storage connection string from file {0}", filePath);
-            return File.ReadAllText(filePath);
-        }
-
-        static string ConnectionStringFromEnvironmentVariable(string environmentVariableName)
-        {
-            var value = Environment.GetEnvironmentVariable(environmentVariableName);
-
-            if (value == null)
-            {
-                Console.WriteLine("Could not find env variable {0}", environmentVariableName);
-                return null;
-            }
-
-            Console.WriteLine("Using Azure Storage connection string from env variable {0}", environmentVariableName);
-
-            return value;
-        }
-
-        static string Throw(string message)
-        {
-            throw new ConfigurationErrorsException(message);
         }
 
         public void CleanUp()