+++ CHANGELOG.md
@@ -960,6 +960,9 @@
 # 0.99.60
 
 * Changed data type of the SQL Server saga storage `[data]` column to `VARBINARY(MAX)` because `NVARCHAR(MAX)` was extremely slow!
+* Added Azure blobs-based data bus storage
+
+
 
 ---